FBTweak.xcodeproj: Stop copying header files.
Before this commmit, the project copied header files to the include
directory after build, which eliminates the need to add header file
paths in projects that use this project. This, however, has some
unwanted side-effects, like the include directory not being cleaned
when clean is performed, and more subtle bugs that causes new header
files not being copied.

Futhermore, since we specify header search path for all libraries, the
both the original header files and the files manually copied by the
project are in the header search path. This may lead to problems when
the copy did not happen for some reason.