Tweaks: Ignore persisted value if its type differs from default value's.
If a persistent tweak was created with a default value one type, then
its current value was set, and in a subsequent run someone changes the
type of the default value to another type, the current value and the
default value have inconsistent types.
To solve this, current value is loaded only if its type agrees with the
type of the default value.

Note: This won't work if a tweak is created using `FBInlineTweak` or
any of the inline macros, because these tweaks create an
`FBPersistentTweak` once upon program execution.