FBTweak: Archive value after it passed possibleValues validation.
Until now, if `currentValue` was set to a persistent tweak, if that
value wasn't one of the possible values, `currentValue` would be set to
`defaultValue` instead. However, the persisted value would still be the
value that was set, meaning, a value that isn't one of the possible
values. This commit changes it by persisting the value after it passed
the possible values validation.