FBTweak: Ignore persisted value if its class isn't the default value's.
There can be a mismatch between a tweak's default value and its current
value, if the default value was changed at a later point for example. To
fix this discrepancy, the persisted value is loaded as the tweak's
current value only if if it has the same kind of class as the default
value's class.